Fixing compilation issues

+++ Mines Courses/CSCI_440/homework4_Carson_Stevens/homework4_Carson_Stevens.cu
@@ -11,7 +11,7 @@ using namespace std;
 
 __global__ void scan_with_addition(float *g_idata, float *g_odata, int n)
 {
-    extern __shared__ float temp[]; // allocated on invocation
+    extern __shared__ int temp[]; // allocated on invocation
     int thid = threadIdx.x;
     int pout = 0, pin = 1;
     // load input into shared memory.
